requirements.txt
json==2.0.9
numpy==1.19.2
pillow==8.0.0
tqdm==4.55.0
pytorch==1.7.1
protobuf==3.15.1
torchvision==0.8.2